+++ library/src/main/java/io/sfrei/tracksearch/tracks/TrackList.java
@@ -39,4 +39,6 @@ public interface TrackList<T extends Track> {
 
     String getQueryParam();
 
+    TrackList<T> next();
+
 }